updated bounty routes, untested
