+++ controller/index.js
@@ -5,5 +5,6 @@ const taskRoutes = require('./taskRoutes.js');
 
 router.use('/task', taskRoutes);
 router.use('/task/subtask', subtaskRoutes);
+router.use('/bounty', bountyRoutes);
 
 module.exports = router;
