+++ model/bountyModel.js
@@ -6,13 +6,10 @@ const Bounty = sequelize.define('Task', {
   id: {
     type: DataTypes.UUID,
     defaultValue: uuidv4, // Assign the UUIDv4 function as the default value
+    allowNull: false,
     primaryKey: true,
   },
-  name: {
-    type: DataTypes.STRING,
-    allowNull: false,
-  },
-  bubbleBucks: {
+  price: {
     type: DataTypes.INTEGER,
     allowNull: false,
   },
